% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_batch_FD.R
\name{run_batch_FD}
\alias{run_batch_FD}
\title{run_batch_FD}
\usage{
run_batch_FD(inputDir, outputDir, FDargs = c("-RTwindow 20", "-minRTspan 3"))
}
\arguments{
\item{FDargs}{}
}
\value{

}
\description{
run_batch_FD
}
\examples{

}
